REV-10787: Remove Formatters.
