+++ composer.json
@@ -3,7 +3,7 @@
     "description": "Laravel report tool",
     "type": "library",
     "require": {
-        "php": "^8.1",
+        "php": "^8.0",
         "laravel/framework": "8.x|9.x"
     },
     "license": "MIT",
